refactor: extract CollaboratorSelect to shared component, reuse in IDP modal
Extract CollaboratorSelect from EvaluationsTab into its own file.
Replace native <select> in DevelopmentPlanFormModal with the custom
dropdown (search, icons, consistent styling). Fix em dashes to parentheses.
